fix: apply cargo fmt + 修复 mount_windows 类型转换
CI 在 push #a33ed7b 之后三个 job（rustfmt / clippy / test-windows）失败。
本地代码基于旧版 rustfmt 风格且含一个生产 bug，统一修复。

修复 1：cargo fmt --all
  - inode.rs (5 行)
  - mount.rs (22 行变化)
  - mount_linux.rs (4 行变化)
  - server/mod.rs (4 行变化)
  - server/webdav.rs (29 行变化)
  触发行：长行换行 / 关联参数重新对齐

修复 2：mount_windows.rs:958 测试中 FspCleanupDelete 缺 as u32
  报错：error[E0308] mismatched types: expected `u32`, found `i32`
  现状：mount.rs:137 / mount.rs:953-961 处的同名字段都已用 `as u32` 转换，
  此处是测试用例里漏写的 cast。生产代码 file.rs 上 path:140 期望 u32。
  修复：FspCleanupDelete → FspCleanupDelete as u32

CI 现在应该：
  ✓ cargo fmt --all -- --check
  ✓ Windows test compile
  待 ci.yml 系统依赖修复后通过 linux/clippy

+++ crates/rs-f4ss-core/src/mount_windows.rs
@@ -955,7 +955,7 @@ mod tests {
         let ctx = FileContext::new_test("/delete.txt", false, false);
         assert_eq!(ctx.close_action(), CloseAction::Release);
 
-        ctx.record_cleanup_result(FspCleanupDelete, true);
+        ctx.record_cleanup_result(FspCleanupDelete as u32, true);
 
         assert!(ctx.delete_requested());
         assert_eq!(ctx.close_action(), CloseAction::Discard);
